Move envvar to specific package
